Add files via upload
This code will take a file as input which will contain the text which need to be masked.The output will be written on screen.It can be tweaked to generate the output in a file. Data that will be masked are 
1)Name
2)Place
3)Organization
4)Phone Number
5)Email Id
6)IP address

work is going on to add Date also in the to be masked list. Feel free to contribute.